+++ src/services/static-config.ts
@@ -139,6 +139,7 @@ class StaticConfig {
       // viewer settings
       high_res_cond_default: 'zoom',
       livephoto_autoplay: true,
+      livephoto_loop: false,
       sidebar_filepath: false,
       metadata_in_slideshow: false,