+++ src/typings/config.d.ts
@@ -26,6 +26,7 @@ declare module '@typings' {
     // viewer settings
     high_res_cond_default: HighResCond;
     livephoto_autoplay: boolean;
+    livephoto_loop: boolean;
     sidebar_filepath: boolean;
     metadata_in_slideshow: boolean;